% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lempel_ziv.R
\name{measure.kolmogorov}
\alias{measure.kolmogorov}
\title{Kolmogorov Complexity}
\usage{
measure.kolmogorov(time_serie)
}
\arguments{
\item{time_serie}{A time serie, object of type ts.}
}
\value{
The value of the complexity measure: Kolmogorov.
}
\description{
Based on http://www.mathworks.com/matlabcentral/fileexchange/6886-kolmogorov-complexity
}
