% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtosis_measures.R
\name{measure.kurtosis}
\alias{measure.kurtosis}
\title{Estimation of Kurtosis}
\usage{
measure.kurtosis(time_serie)
}
\arguments{
\item{time_serie}{A time serie, object of type ts.}
}
\value{
The value of the Kurtosis measure
}
\description{
Using https://cran.r-project.org/web/packages/e1071/e1071.pdf
}
