% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_entropy.R
\name{measure.shannonEntropy}
\alias{measure.shannonEntropy}
\title{Estimation of Shannon Entropy}
\usage{
measure.shannonEntropy(time_serie, method)
}
\arguments{
\item{time_serie}{A time serie, object of type ts.}

\item{method}{how to calculate shannon entropy: "SG", "CS", "ML"}
}
\value{
The value of the complexity measure: Estimation of Entropy.
}
\description{
Using http://cran.r-project.org/web/packages/entropy/entropy.pdf
Bins https://www.jmlr.org/papers/volume13/brown12a/brown12a.pdf
}
