% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness_measures.R
\name{measure.skewness}
\alias{measure.skewness}
\title{Estimation of Skewness}
\usage{
measure.skewness(time_serie)
}
\arguments{
\item{time_serie}{A time serie, object of type ts.}
}
\value{
The value of the Skewness measure
}
\description{
Using https://cran.r-project.org/web/packages/e1071/e1071.pdf
}
