% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{measure.spectral_entropy}
\alias{measure.spectral_entropy}
\title{Spectral Entropy}
\usage{
measure.spectral_entropy(y)
}
\arguments{
\item{y}{input time series data}
}
\value{
spectral entropy
}
\description{
Calculation of the value of the spectral entropy.
Based on: https://rdrr.io/cran/ForeCA/man/spectral_entropy.html
}
