% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsfeatures.R
\name{tsfeaturesH}
\alias{tsfeaturesH}
\title{Calculate some interesting features of the time series based on the tsfeatures package}
\usage{
tsfeaturesH(ts, fast)
}
\arguments{
\item{ts}{imput time series.}
}
\description{
Using package tsfeatures (https://www.rdocumentation.org/packages/tsfeatures/versions/1.0.1)
Methods for extracting various features from time series data. The features provided are those
from Hyndman, Wang and Laptev (2013) <doi:10.1109/ICDMW.2015.104>, Kang, Hyndman and
Smith-Miles (2017) <doi:10.1016/j.ijforecast.2016.09.004> and from Fulcher,
Little and Jones (2013) <doi:10.1098/rsif.2013.0048>.
}
